# swagger.yaml
openapi: 3.0.3
info:
  title: Tridentum Care API
  description: "Documentazione API per Tridentum Care, piattaforma di coordinamento dei volontari della città di Trento."
  version: 1.0.0
servers:
  - url: https://tridentum-care.onrender.com/api/v1
    description: "Sviluppo locale"
...
paths:
...
  /volunteers/me:
    get:
      tags:
        - Volunteers
      summary: "Ottieni il profilo del volontario"
      description: "Restituisce il profilo completo del volontario autenticato, inclusi competenze, punti e coupon acquisiti."
      security:
        - bearerAuth: []
      x-role-required: volunteer
      responses:
        '200':
          description: "Profilo del volontario recuperato con successo"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: "Non autorizzato"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: "Proibito (ruolo sbagliato)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Volontario non trovato"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Errore interno del server"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Volunteers
      summary: "Aggiorna il profilo del volontario"
      description: "Aggiorna informazioni e competenze personali. authLvl non può essere modificato qui."
      security:
        - bearerAuth: []
      x-role-required: volunteer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VolunteerProfileUpdate'
      responses:
        '200':
          description: "Profilo aggiornato con successo"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: "Errore di convalida (ad esempio, età < 18)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Non autorizzato"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: "Vietato"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Utente non trovato"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Errore interno del server"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Volunteers
      summary: "Elimina il profilo del volontario"
      description: "Cancellazione conforme al GDPR. Reimposta le richieste \"In Corso\", anonimizza gli altri ed elimina il documento utente."
      security:
        - bearerAuth: []
      x-role-required: volunteer
      responses:
        '200':
          description: "Profilo eliminato con successo"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: "Non autorizzato"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: "Vietato"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Utente non trovato"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Errore interno del server"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /volunteers/coupons:
    get:
      tags:
        - Volunteers
      summary: "Ottieni i coupon del negozio disponibili"
      description: "Restituisce tutti i coupon che non sono ancora scaduti."
      security:
        - bearerAuth: []
      x-role-required: volunteer
      responses:
        '200':
          description: "Elenco dei coupon attivi"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Coupon'
        '401':
          description: "Non autorizzato"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: "Vietato"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Errore interno del server"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /volunteers/coupons/redemptions:
    post:
      tags:
        - Volunteers
      summary: "Riscatta un buono"
      description: "Detrae punti dal volontario e garantisce un codice coupon univoco."
      security:
        - bearerAuth: []
      x-role-required: volunteer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - couponId
              properties:
                couponId:
                  type: string
                  description: "L'ID del coupon da riscattare"
      responses:
        '200':
          description: "Coupon riscattato con successo"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  newPoints:
                    type: integer
        '400':
          description: "Punti insufficienti o couponId mancante"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Non autorizzato"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: "Vietato"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Coupon o utente non trovato"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Errore interno del server"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /volunteers/me/tasks:
    get:
      tags:
        - Volunteers
      summary: "Ottieni i compiti assegnati al volontario"
      description: "Restituisce tutte le richieste attualmente assegnate al volontario autenticato con stato \"In Corso\", ordinate per data di creazione decrescente."
      security:
        - bearerAuth: []
      x-role-required: volunteer
      responses:
        '200':
          description: "Elenco delle attività assegnate recuperate correttamente"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Request'
        '401':
          description: "Non autorizzato"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: "Proibito (ruolo sbagliato)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Volontario non trovato"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Errore interno del server"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        id:
          type: string
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [requester, volunteer, partner]
        phone:
          type: string
        address:
          type: string
        createdAt:
          type: string
          format: date-time
        authLvl:
          type: integer
          enum: [0, 1, 2]
        points:
          type: integer
        skills:
          type: array
          items:
            type: string
        coupons:
          type: array
          items:
            $ref: '#/components/schemas/OwnedCoupon'
        age:
          type: integer
        gender:
          type: string
        license:
          type: string
        isSuspended:
          type: boolean
        suspendedUntil:
          type: string
          format: date-time
    Request:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        serviceType:
          type: string
        category:
          type: string
        title:
          type: string
        location:
          type: string
        address:
          type: string
        date:
          type: string
        time:
          type: string
        dateTime:
          type: string
        notes:
          type: string
        description:
          type: string
        points:
          type: integer
        status:
          type: string
          enum:
            - In Attesa di Volontario
            - In Corso
            - Completata
            - Annullata
        volunteerId:
          type: string
          nullable: true
        requesterName:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
    Coupon:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        pointsCost:
          type: integer
        expirationDate:
          type: string
        partnerId:
          type: string
        companyName:
          type: string
          description: "Nome dell'azienda partner che offre il coupon"
    OwnedCoupon:
      type: object
      properties:
        name:
          type: string
        cost:
          type: integer
        code:
          type: string
        acquiredAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
...
    VolunteerProfileUpdate:
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        address:
          type: string
        phone:
          type: string
        skills:
          type: array
          items:
            type: string
        age:
          type: integer
          minimum: 18
        license:
          type: string
        gender:
          type: string